s3_client: Return data in an 8-byte aligned slice
The data returned by the AWS S3 client does not have any alignment
guarantees. In order to reinterpret the data as an array of numbers with
a higher alignment than 1, we need to return the data in Bytes object in
which the underlying data has a higher alignment.  For now we're
hard-coding an alignment of 8 bytes, although this should depend on the
data type, and potentially whether there are any SIMD requirements.

The current method is rather inefficient, involving copying the data at
least twice. This is functional, but should be revisited.