Allow using Rayon to execute CPU-bound tasks
Once data has been downloaded from S3, the remaining tasks in the
pipeline (filters, decompression, numerical operations) are generally
CPU-bound. Tokio is an asynchronous runtime and is not designed to
handle long-lived CPU-bound tasks. If too many CPU-bound tasks are
executing, this will starve other asynchronous tasks, and prevent them
from making progress.

Rayon is a data-parallelism library that uses a fixed-size pool of
worker threads to execute tasks. The tokio-rayon crate provides an
adapter between Tokio and Rayon that allows us to schedule tasks to a
Rayon thread pool and wait for them to complete.

This change adds a --use-rayon CLI flag that causes reductionist to use
Rayon thread pool to execute CPU-bound tasks. The Rayon thread pool is
configured to allow one CPU core free for the Tokio runtime to use.