Fix trim trailing slash middleware
The axum docs point out that any middleware that should affect routing must wrap
the Router. In 0.6.12 we rely on a bug in axum routing that allows the
trailing slash middleware to work on nested paths (but fails on root
paths). In 0.6.13 onwards this bug is no longer present, so we lose the
trailing slash removal.

This change fixes the issue.

https://github.com/tokio-rs/axum/issues/1978
https://docs.rs/axum/0.6.18/axum/middleware/index.html#rewriting-request-uri-in-middleware