Refactor into a library with a minimal binary
This change refactors the code from a binary project into a library.
The main.rs file is now an included binary that uses the
s3_active_storage library.

This opens up the potential for extension by building a different crate
that uses this crate, and adds or modifies its behaviour.