Add support for Gzip and Zlib compression
This allows the S3 active storage server to work with data that was
compressed using one of these algorithms before it was written to the
object store.

Compression is configured using the optional `compression` field in the
API JSON request data, and if present should be set to "gzip" or "zlib".

We are using the standard flate2 library for decompression. There may be
more performant options to consider in future, but this works well as a
first pass, and changing in future will not affect the API.