Support byte shuffle filter
Both Zarr and HDF5 provide a byte shuffle filter which can be used to
improve compression ratios of certain data patterns.

There does not appear to be a Rust implementation of the byte shuffle
filter, so this change includes one.

It also adds a "filters" parameter to the API request data which accepts
a list of filter definitions. Each filter definition is a dict with an
"id" field and optional parameter fields specific to each filter. The
shuffle filter has one parameter, "element_size".