Perform size validation after decompression
When data is compressed, the size parameter refers to the size of the
compressed data. Typically this is not equal to the size of the
uncompressed data, so we can't validate it against the data type size.

This change skips initial size/dtype validation when compression is
applied, instead performing it once the data has been decompressed.

It also adds an additional validation that the size matches the shape,
when a shape has been specified.