Rebrand to Reductionist
The S3 Active Storage Server is a bit of a mouthful, too similar to
PyActiveStorage [1], and unnecessarily tied to S3.

From Oxford Languages:

  reductionist:
    noun
      a person who analyses and describes a complex phenomenon in terms
      of its simple or fundamental constituents.
    adjective
      analysing and describing a complex phenomenon in terms of its
      simple or fundamental constituents.

But without thinking about it too hard, this project provides a service
that performs reductions, i.e. a reductionist.

The GitHub repository will be renamed to stackhpc/reductionist-rs.

[1] https://github.com/valeriupredoi/PyActiveStorage