Missing: perform reduction and count in one iteration
This change modifies the min, max and sum operations to perform the
reduction and count in a single iteration. This improved performance
significantly (50%+) based on benchmarks.

This change also adds various unit tests for missing data, as well as
some exploration of infinite and NaN values in data. There seems to be
some inconsistent behaviour here, and more work is required to determine
what behaviour is expected.