Add support for byte order
Adds a new optional "byte_order" parameter to API request data that can
be used to specify the byte order of the array data. It accepts values
of "big" and "little".

If the byte order of the data does not match the native byte order of
the server running Reductionist, the data's byte order is reversed prior
to performing the operation.

A new "x-activestorage-byte-order" header is returned that describes the
byte order of the response payload. This is always the native byte order
of the server running Reductionist.