tracing: add more instrumentation, optionally send to jaeger
Tracing allows us to gain insight into the performance of reductionist.
Jaeger provides a UI to view traces, and sending traces is optional and
disabled by default. Enable using --enable-jaeger CLI argument.
Currently only supports sending to localhost.

A slight inconvenience is that if a span is open across an 'await' call,
we need to instrument the future using tracing::instrument() to ensure
the span is entered when we finish waiting.