s3_client: Share S3Client objects
The AWS SDK recommends to share aws_sdk_s3::Client objects due to
significant setup costs and the use of connection pooling. This change
adds an S3ClientMap object that stores shared S3Client objects for each
URL/username/password combination.

Currently these clients are never removed, so the map may grow
unbounded. However, in practice we'll typically be using a single object
store with a limited number of credentials, so this should not be an
immediate problem.