Support unauthenticated access to public buckets
Use of a basic authentication header in requests to the Reductionist API
is now optional. If basic auth credentials are provided, they will be
used as S3 access/secret keys as before. Otherwise, requests to S3 will
be unauthenticated. Attempts to access a private bucket without
authentication will return a 401.